Added intro-app's `App.js` to my-app's `app/index.tsx` + removed intro-app
